Flytt batch-jobb til admin og legg på mer logging og throttling
